requirements.txt
opencv-python
mediapipe
numpy
torch
torchvision
scikit-learn
matplotlib
psutil
tqdm
pyyaml
